+++ pmd-apex/src/test/java/net/sourceforge/pmd/lang/apex/rule/bestpractices/ApexUnitTestClassShouldHaveRunAsTest.java
@@ -2,10 +2,10 @@
  * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
  */
 
-package net.sourceforge.pmd.lang.java.rule.errorprone;
+package net.sourceforge.pmd.lang.apex.rule.bestpractices;
 
 import net.sourceforge.pmd.testframework.PmdRuleTst;
 
-public class BeanMembersShouldSerializeTest extends PmdRuleTst {
+public class ApexUnitTestClassShouldHaveRunAsTest extends PmdRuleTst {
     // no additional unit tests
 }