+++ pmd-core/src/main/java/net/sourceforge/pmd/lang/document/NioTextFile.java
@@ -17,8 +17,8 @@ import net.sourceforge.pmd.internal.util.AssertionUtil;
 import net.sourceforge.pmd.lang.LanguageVersion;
 import net.sourceforge.pmd.util.IOUtil;
 import net.sourceforge.pmd.util.datasource.DataSource;
-import net.sourceforge.pmd.util.datasource.FileDataSource;
 import net.sourceforge.pmd.util.datasource.internal.LanguageAwareDataSource;
+import net.sourceforge.pmd.util.datasource.internal.PathDataSource;
 
 /**
  * A {@link TextFile} backed by a file in some {@link FileSystem}.
@@ -41,7 +41,8 @@ class NioTextFile implements TextFile {
         this.path = path;
         this.charset = charset;
         this.languageVersion = languageVersion;
-        this.pathId = path.toAbsolutePath().toString();
+        // using the URI here, that handles files inside zip archives automatically (schema "jar:file:...!/path/inside/zip")
+        this.pathId = path.toUri().toString();
     }
 
     @Override
@@ -74,7 +75,7 @@ class NioTextFile implements TextFile {
 
     @Override
     public DataSource toDataSourceCompat() {
-        return new LanguageAwareDataSource(new FileDataSource(path.toFile()), languageVersion);
+        return new LanguageAwareDataSource(new PathDataSource(path), languageVersion);
     }
 
     @Override