+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/rule/AbstractJUnitRule.java
@@ -4,7 +4,11 @@
 
 package net.sourceforge.pmd.lang.java.rule;
 
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 
 import net.sourceforge.pmd.RuleContext;
 import net.sourceforge.pmd.annotation.InternalApi;
@@ -30,6 +34,15 @@ public abstract class AbstractJUnitRule extends AbstractJavaRule {
     protected static final String JUNIT3_CLASS_NAME = "junit.framework.TestCase";
     protected static final String JUNIT4_CLASS_NAME = "org.junit.Test";
     protected static final String JUNIT5_CLASS_NAME = "org.junit.jupiter.api.Test";
+    private static final String JUNIT5_NESTED = "org.junit.jupiter.api.Nested";
+    private static final Set<String> JUNIT5_TEST_ANNOTATIONS = new HashSet<>(Arrays.asList(
+            JUNIT5_CLASS_NAME,
+            "org.junit.jupiter.api.RepeatedTest",
+            "org.junit.jupiter.api.TestFactory",
+            "org.junit.jupiter.api.TestTemplate",
+            "org.junit.jupiter.params.ParameterizedTest"));
+
+    private static final String TESTNG_ANNOTATION = "org.testng.annotations.Test";
 
     protected boolean isJUnit3Class;
     protected boolean isJUnit4Class;
@@ -92,13 +105,13 @@ public abstract class AbstractJUnitRule extends AbstractJavaRule {
 
         if (isJUnit4Class && isJUnit5Class) {
             isJUnit4Class &= hasImports(node, JUNIT4_CLASS_NAME);
-            isJUnit5Class &= hasImports(node, JUNIT5_CLASS_NAME);
+            isJUnit5Class &= hasImports(node, JUNIT5_TEST_ANNOTATIONS);
         }
     }
 
     public static boolean isTestClass(ASTClassOrInterfaceBody node) {
-        return !isAbstractClass(node)
-                && (isTestClassJUnit3(node) || isTestClassJUnit4(node) || isTestClassJUnit5(node));
+        return !isAbstractClass(node) && node.getParent() instanceof ASTClassOrInterfaceDeclaration
+                && (isTestClassJUnit3(node) || isTestClassJUnit4(node) || isTestClassJUnit5(node) || isTestClassTestNg(node));
     }
 
     private static boolean isAbstractClass(ASTClassOrInterfaceBody node) {
@@ -131,11 +144,19 @@ public abstract class AbstractJUnitRule extends AbstractJavaRule {
         Node parent = node.getParent();
         if (parent instanceof TypeNode) {
             TypeNode type = (TypeNode) parent;
-            return isJUnit5Class(type) && hasImports(type, JUNIT5_CLASS_NAME);
+            return isJUnit5Class(type) && hasImports(type, JUNIT5_TEST_ANNOTATIONS);
         }
         return false;
     }
 
+    private static boolean isTestClassTestNg(ASTClassOrInterfaceBody node) {
+        Node parent = node.getParent();
+        if (parent instanceof TypeNode) {
+            TypeNode type = (TypeNode) parent;
+            return doesNodeContainJUnitAnnotation(type, TESTNG_ANNOTATION);
+        }
+        return false;
+    }
 
     public static boolean isTestMethod(ASTMethodDeclaration method) {
         if (method.isAbstract() || method.isNative() || method.isStatic()) {
@@ -145,7 +166,12 @@ public abstract class AbstractJUnitRule extends AbstractJavaRule {
         ASTClassOrInterfaceBody type = method.getFirstParentOfType(ASTClassOrInterfaceBody.class);
         return isTestClassJUnit3(type) && method.isPublic() && method.isVoid() && method.getName().startsWith("test")
                 || isTestClassJUnit4(type) && method.isPublic() && doesNodeContainJUnitAnnotation(method.getParent(), JUNIT4_CLASS_NAME)
-                || isTestClassJUnit5(type) && doesNodeContainJUnitAnnotation(method.getParent(), JUNIT5_CLASS_NAME);
+                || isTestClassJUnit5(type) && doesNodeContainJUnitAnnotation(method.getParent(), JUNIT5_TEST_ANNOTATIONS)
+                || doesNodeContainJUnitAnnotation(method.getParent(), TESTNG_ANNOTATION);
+    }
+
+    public static boolean isJUnit5NestedClass(ASTClassOrInterfaceBody innerClassDecl) {
+        return doesNodeContainJUnitAnnotation((JavaNode) innerClassDecl.getNthParent(2), JUNIT5_NESTED);
     }
 
     public boolean isJUnitMethod(ASTMethodDeclaration method, Object data) {
@@ -170,7 +196,7 @@ public abstract class AbstractJUnitRule extends AbstractJavaRule {
     }
 
     private boolean isJUnit5Method(ASTMethodDeclaration method) {
-        return isJUnit5Class && doesNodeContainJUnitAnnotation(method.getParent(), JUNIT5_CLASS_NAME);
+        return isJUnit5Class && doesNodeContainJUnitAnnotation(method.getParent(), JUNIT5_TEST_ANNOTATIONS);
     }
 
     private boolean isJUnit3Method(ASTMethodDeclaration method) {
@@ -186,34 +212,72 @@ public abstract class AbstractJUnitRule extends AbstractJavaRule {
     }
 
     private static boolean isJUnit5Class(JavaNode node) {
-        return doesNodeContainJUnitAnnotation(node, JUNIT5_CLASS_NAME);
+        return doesNodeContainJUnitAnnotation(node, JUNIT5_TEST_ANNOTATIONS);
     }
 
     private static boolean doesNodeContainJUnitAnnotation(JavaNode node, String annotationTypeClassName) {
+        return doesNodeContainJUnitAnnotation(node, Collections.singleton(annotationTypeClassName));
+    }
+
+    private static boolean doesNodeContainJUnitAnnotation(JavaNode node, Set<String> annotationTypeClassNames) {
         List<ASTAnnotation> annotations = node.findDescendantsOfType(ASTAnnotation.class);
         for (ASTAnnotation annotation : annotations) {
             Node annotationTypeNode = annotation.getChild(0);
             TypeNode annotationType = (TypeNode) annotationTypeNode;
             if (annotationType.getType() == null) {
                 ASTName name = annotationTypeNode.getFirstChildOfType(ASTName.class);
-                if (name != null && (name.hasImageEqualTo("Test") || name.hasImageEqualTo(annotationTypeClassName))) {
-                    return true;
+                if (name != null) {
+                    for (String annotationTypeName : annotationTypeClassNames) {
+                        if (areNamesEqual(name, annotationTypeName)) {
+                            return true;
+                        }
+                    }
                 }
-            } else if (TypeTestUtil.isA(annotationTypeClassName, annotationType)) {
-                return true;
+            } else {
+                for (String annotationTypeClassName : annotationTypeClassNames) {
+                    if (TypeTestUtil.isA(annotationTypeClassName, annotationType)) {
+                        return true;
+                    }
+                }
             }
         }
         return false;
     }
 
     private static boolean hasImports(JavaNode node, String className) {
+        return hasImports(node, Collections.singleton(className));
+    }
+
+    private static boolean hasImports(JavaNode node, Set<String> classNames) {
         List<ASTImportDeclaration> imports = node.getRoot().findDescendantsOfType(ASTImportDeclaration.class);
         for (ASTImportDeclaration importDeclaration : imports) {
             ASTName name = importDeclaration.getFirstChildOfType(ASTName.class);
-            if (name != null && name.hasImageEqualTo(className)) {
+            // imports are always fully qualified
+            if (name != null && classNames.contains(name.getImage())) {
                 return true;
             }
         }
         return false;
+    }
+
+    /**
+     * Compares the name against the given fully qualified name. If a direct comparison doesn't
+     * work, try to compare as a simple name.
+     *
+     * <p>Note: This method is only used, if the auxclasspath is incomplete.</p>
+     */
+    private static boolean areNamesEqual(ASTName node, String fullyQualifiedAnnotationName) {
+        String simpleName = node.getImage();
+        if (simpleName.equals(fullyQualifiedAnnotationName)) {
+            return true;
+        }
+
+        String simpleAnnotationName = fullyQualifiedAnnotationName;
+        int lastDot = fullyQualifiedAnnotationName.lastIndexOf('.');
+        if (lastDot != -1) {
+            simpleAnnotationName = fullyQualifiedAnnotationName.substring(lastDot + 1);
+        }
+        // when comparing by simple name, double check whether we have a import to avoid false positives
+        return simpleName.equals(simpleAnnotationName) && hasImports(node, fullyQualifiedAnnotationName);
     }
 }