+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/typeresolution/typeinference/TypeInferenceResolver.java
@@ -234,7 +234,9 @@ public final class TypeInferenceResolver {
             for (Variable var : variablesToResolve) {
                 List<JavaTypeDefinition> lowerBounds = getLowerBoundsOf(var, bounds);
                 // TODO: should call incorporation
-                instantiations.put(var, lub(lowerBounds));
+                if (!lowerBounds.isEmpty()) {
+                    instantiations.put(var, lub(lowerBounds));
+                }
             }
 
             uninstantiatedVariables.removeAll(variablesToResolve);