+++ pmd-java/src/test/java/net/sourceforge/pmd/lang/java/rule/errorprone/constructorcallsoverridablemethod/AbstractThing.java
@@ -1,0 +1,19 @@
+/*
+ * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
+ */
+
+package net.sourceforge.pmd.lang.java.rule.errorprone.constructorcallsoverridablemethod;
+
+public abstract class AbstractThing implements Thing {
+    protected AbstractThing(Thing original) {
+        setName(original.getName());
+    }
+
+    @Override
+    public void setName(String name) { }
+
+    @Override
+    public String getName() {
+        return "";
+    }
+}